project_mercury/rtm/src/test/java/com/smartsparrow/rtm/subscription/courseware/evaluableset/EvaluableSetRTMConsumableTest.java
package com.smartsparrow.rtm.subscription.courseware.evaluableset;

import static com.smartsparrow.courseware.data.CoursewareElementType.ACTIVITY;
import static com.smartsparrow.courseware.data.EvaluationMode.DEFAULT;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;

import java.util.UUID;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import com.smartsparrow.rtm.message.RTMClientContext;
import com.smartsparrow.rtm.subscription.courseware.ActivityRTMSubscription;
import com.smartsparrow.rtm.subscription.courseware.message.EvaluableSetBroadcastMessage;
import com.smartsparrow.util.UUIDs;

class EvaluableSetRTMConsumableTest {

    @Mock
    private RTMClientContext rtmClientContext;

    private static final UUID activityId = UUIDs.timeBased();

    @BeforeEach
    void setUp() {
        MockitoAnnotations.openMocks(this);
    }

    @Test
    void create() {
        EvaluableSetBroadcastMessage message = new EvaluableSetBroadcastMessage(activityId, activityId, ACTIVITY, DEFAULT);
        EvaluableSetRTMConsumable consumable = new EvaluableSetRTMConsumable(rtmClientContext, message);

        assertEquals(rtmClientContext, consumable.getRTMClientContext());
        assertEquals(new EvaluableSetRTMEvent().getName(), consumable.getRTMEvent().getName());
        assertEquals(String.format("author.activity/%s/%s",
                                   activityId,
                                   new EvaluableSetRTMEventDecoratorImpl(new EvaluableSetRTMEvent()).getName(
                                           ACTIVITY)), consumable.getName());
        Assertions.assertEquals(ActivityRTMSubscription.NAME(activityId), consumable.getSubscriptionName());
        assertNotNull(consumable.getName());
    }

}
